# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_the-monkeys_freerangenotify_internal_domain_resourcelink.ImportRequest:
    properties:
      resources:
        items:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_resourcelink.ResourceType'
        minItems: 1
        type: array
      source_app_id:
        type: string
    required:
    - resources
    - source_app_id
    type: object
  github_com_the-monkeys_freerangenotify_internal_domain_resourcelink.ResourceType:
    enum:
    - users
    - templates
    - workflows
    - digest_rules
    - topics
    - providers
    type: string
    x-enum-varnames:
    - TypeUser
    - TypeTemplate
    - TypeWorkflow
    - TypeDigest
    - TypeTopic
    - TypeProvider
...
host: localhost:8080
info:
  contact:
    email: support@freerangenotify.com
    name: API Support
  description: High-performance notification service with multi-channel delivery support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FreeRangeNotify API
  version: 0.1.9-alpha
paths:
...
  /v1/apps/{id}/import:
    post:
      consumes:
      - application/json
      description: Create cross-app resource links to share users, templates, workflows,
        etc. between applications
      parameters:
      - description: Target Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Import request specifying source app and resource types
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_resourcelink.ImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Import resources from another app
      tags:
      - Resource Linking
  /v1/apps/{id}/links:
    delete:
      description: Remove all cross-app resource links for an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove all resource links
      tags:
      - Resource Linking
    get:
      description: List all cross-app resource links for an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by resource type
        in: query
        name: resource_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List resource links
      tags:
      - Resource Linking
  /v1/apps/{id}/links/{link_id}:
    delete:
      description: Remove a specific cross-app resource link
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Link ID
        in: path
        name: link_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove a resource link
      tags:
      - Resource Linking
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Application API key for tenant-facing endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and your API key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"